application/requirements.txt
google-cloud-aiplatform
google-cloud-discoveryengine
google-cloud-dlp==3.16.0
google-cloud-storage==2.15.0
google-cloud-logging==3.9.0
vertexai==1.70.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
fastapi-microsoft-identity==0.1.6
google-auth==2.28.1
requests
sqlmodel==0.0.21
psycopg2-binary==2.9.9
openpyxl==3.1.5
xlrd==2.0.1
tabulate==0.9.0
pandas==2.2.1
Faker==30.3.0
gender-guesser==0.4.0
